+++ HISTORY.rst
@@ -3,6 +3,7 @@ History
 =======
 
 * Updated client/server comms & drivers to reuse response comms
+* Added additional patterns to ForkComm
 
 1.6.1 (2021-05-18) Minor Bug Fix
 ------------------